FROMLIST: block: genhd: Add disk/partition specific uevent callbacks for partition info
This patch has been carried in the Android tree for quite some
time and is one of the few patches required to get a mainline
kernel up and running with an exsiting Android userspace. So I
wanted to submit it for review and consideration if it should
be merged.

For disk devices, a new uevent parameter 'NPARTS' specifies the number
of partitions detected by the kernel. Partition devices get 'PARTN' which
specifies the partitions index in the table, and 'PARTNAME', which
specifies PARTNAME specifices the partition name of a partition device

Android's userspace uses this for creating device node links from the
partition name and number: ie:
	/dev/block/platform/soc/by-name/system
or
	/dev/block/platform/soc/by-num/p1

One can see its usage here:
https://android.googlesource.com/platform/system/core/+/master/init/devices.cpp#355
and
https://android.googlesource.com/platform/system/core/+/master/init/devices.cpp#494

Thoughts and feedback would be greatly appreciated!

Cc: Jens Axboe <axboe@kernel.dk>
Cc: Rom Lemarchand <romlem@google.com>
Cc: Android Kernel Team <kernel-team@android.com>
Signed-off-by: Dima Zavin <dima@android.com>
[jstultz: Added more context to commit message & whitespace fix]
Signed-off-by: John Stultz <john.stultz@linaro.org>

Bug: 25923642
Patchset: rk3036 evb bring up

(am from https://patchwork.kernel.org/patch/7466231/)
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>

Change-Id: I14ec24cfd5ec06f31b3705fd18b3f9f25beb30a5