FROMLIST: platform: goldfish: pipe: add devicetree bindings
Add bindings so we don't need to rely on goldfish virtual bus for
probing any more, which means we don't need ARM and MIPS goldfish
board code for instantiating the bus.

In the long term we would like to move towards replacing the Android
pipe with virtio-vsock that is currently under development.

Signed-off-by: Greg Hackmann <ghackmann@google.com>
Signed-off-by: Jin Qian <jinqian@android.com>
Acked-by: Rob Herring <robh@kernel.org>

Bug: 25515960
Patchset: goldfish support

Change-Id: If03054b8739bd929434c8137d3a106847a1e9287
Message-Id: <1449084968-30211-4-git-send-email-jinqian@android.com>
Signed-off-by: Kees Cook <keescook@chromium.org>