FROMLIST: platform: goldfish: pipe: don't log when dropping PIPE_ERROR_AGAIN
On PIPE_ERROR_AGAIN, just stopping in the middle of a transfer and
returning the number of bytes actually handled is the right behavior.

Other errors should be returned on the next read() or write() call.
Continue logging those until we confirm nothing actually relies on the
existing (wrong) behavior of dropping errors on the floor.

Signed-off-by: Greg Hackmann <ghackmann@google.com>
Signed-off-by: Jin Qian <jinqian@android.com>

Bug: 25515960
Patchset: goldfish support

Change-Id: I9cdff243a776018de8dc3ccd75dfb23415dd1aa8
Message-Id: <1449084968-30211-5-git-send-email-jinqian@android.com>
Signed-off-by: Kees Cook <keescook@chromium.org>