FROMLIST: ASoC: rockchip: i2s: add support for grabbing output clock to codec
We need to claim the clock which is driving the codec so that when we
enable clock gating, we continue to clock the codec when needed.
I make this an optional clock since there might be some applications
where we don't need it but can still use the I2S block.

Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Jianqun Xu <jay.xu@rock-chips.com>
Signed-off-by: Caesar Wang <wxt@rock-chips.com>

Bug: 25923642
Patchset: Support the output clock for codec.

(am from https://patchwork.kernel.org/patch/8041001/)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Change-Id: Ib3b44613214ee15f69060bee38b86337afc4d472