UPSTREAM: ARM: dts: rockchip: increase the mclk_fs to 512 for kylin board
If we playback the 8KHz FS audio with the 256 mclk_fs, we need the
mclk = 256 * 8000 = 2.048MHz, the frac div is 594 / 2.048 = 290,
the frac div value 0x00809015 set to the CRU_CLKSEL7_CON will cause
to hang.

We increase the mclk_fs to 512, will get the mclk = 512 * 8000 =
4.096MHz, use 0x01009015 instead of 0x00809015 to work around this
issue. We will keep tracking it.

Signed-off-by: Xing Zheng <zhengxing@rock-chips.com>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>

Bug: 25923642
Patchset: Support to playback 8KHz got audio.
(Note: fixit as the 512 since the upstream is 256)

(cherry picked from git.kernel.org mmind/linux-rockchip.git v4.6-armsoc/dts32
 commit f6bb9d5f30d6986c4fdce1ed5a36088a0c30c544)
Signed-off-by: Caesar Wang <wxt@rock-chips.com>

Change-Id: Id6f5a864f6d6f8f3dbcadd4c03fa5882000711dd